Separate fileInfo structs into its own module
